refactor: change schema
