+++ src/api/database/seeds/createProjects.ts
@@ -1,4 +1,5 @@
 import { prisma } from '../prisma/client.js';
+import { globalOrganization } from './createOrganizations.js';
 
 export const globalProject = '10000000-1000-0000-0000-000000000001';
 
@@ -7,8 +8,9 @@ export const createProjects = async (): Promise<void> => {
     data: [
       {
         id: globalProject,
+        organizationId: globalOrganization,
         slug: 'global',
-        name: 'global-project',
+        name: 'Global Project',
       },
     ],
   });
