+++ src/api/database/seeds/createRoles.ts
@@ -1,5 +1,5 @@
 import { prisma } from '../prisma/client.js';
-import { globalProject } from './createProject.js';
+import { globalProject } from './createProjects.js';
 
 export const adminRole = '10000000-2000-0000-0000-000000000001';
 export const editorRole = '10000000-2000-0000-0000-000000000002';
@@ -14,21 +14,18 @@ export const createRoles = async (): Promise<void> => {
         projectId: globalProject,
         name: 'Administrator',
         description: 'Administrator',
-        adminAccess: true,
       },
       {
         id: editorRole,
         projectId: globalProject,
         name: 'Editor',
         description: 'Editor',
-        adminAccess: false,
       },
       {
         id: guestRole,
         projectId: globalProject,
         name: 'Guest',
         description: 'Guest',
-        adminAccess: false,
       },
     ],
   });
