+++ src/api/database/seeds/createUsers.ts
@@ -2,7 +2,7 @@ import { faker } from '@faker-js/faker';
 import { v4 as uuidV4 } from 'uuid';
 import { oneWayHash } from '../../utilities/oneWayHash.js';
 import { prisma } from '../prisma/client.js';
-import { globalProject } from './createProject.js';
+import { globalProject } from './createProjects.js';
 import { adminRole, editorRole, guestRole } from './createRoles.js';
 
 export const adminUser = '10000000-2000-1000-0000-000000000001';
@@ -18,7 +18,6 @@ export const createUsers = async (): Promise<void> => {
         // Global Project
         {
           id: adminUser,
-          roleId: adminRole,
           name: faker.person.firstName() + ' ' + faker.person.lastName(),
           email: 'admin@collections.dev',
           avatarUrl: faker.internet.avatar(),
@@ -28,7 +27,7 @@ export const createUsers = async (): Promise<void> => {
         },
         {
           id: editorUser,
-          roleId: editorRole,
+
           name: faker.person.firstName() + ' ' + faker.person.lastName(),
           email: 'editor@collections.dev',
           avatarUrl: faker.internet.avatar(),
@@ -38,7 +37,6 @@ export const createUsers = async (): Promise<void> => {
         },
         {
           id: guestUser,
-          roleId: guestRole,
           name: faker.person.firstName() + ' ' + faker.person.lastName(),
           email: 'guest@collections.dev',
           avatarUrl: faker.internet.avatar(),
@@ -55,14 +53,20 @@ export const createUsers = async (): Promise<void> => {
         {
           projectId: globalProject,
           userId: adminUser,
+          roleId: adminRole,
+          isAdmin: true,
         },
         {
           projectId: globalProject,
           userId: editorUser,
+          roleId: editorRole,
+          isAdmin: false,
         },
         {
           projectId: globalProject,
           userId: guestUser,
+          roleId: guestRole,
+          isAdmin: false,
         },
       ],
     });
