+++ src/api/database/seeds/dev.ts
@@ -1,10 +1,12 @@
 import { Output } from '../../../utilities/output.js';
-import { createProjects } from './createProject.js';
+import { createOrganizations } from './createOrganizations.js';
+import { createProjects } from './createProjects.js';
 import { createRoles } from './createRoles.js';
 import { createUsers } from './createUsers.js';
 
 export const seedDev = async (): Promise<void> => {
   try {
+    await createOrganizations();
     await createProjects();
     await createRoles();
     await createUsers();
